livecc/requirements.txt
absl-py==2.2.0
accelerate==1.5.2
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alabaster==0.7.16
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
av>=12.0.0
babel==2.16.0
beautifulsoup4==4.13.3
bidict==0.23.1
bleach==6.2.0
blessed==1.20.0
blinker==1.5
certifi==2024.8.30
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
datasets==3.4.1
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.4
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distlib==0.3.9
distro==1.8.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.19
easydict==1.13
einops==0.8.1
executing==2.2.0
fastapi==0.115.11
fastavro==1.10.0
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gpg==1.18.0
gpustat==1.1.1
gradio==5.26.0
gradio_client==1.9.0
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
hjson==3.1.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib-metadata==6.7.0
iniconfig==2.1.0
insightface==0.7.3
iotop==0.6
ipaddress==1.0.23
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
iso8601==1.0.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonargparse==4.14.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lazy_loader==0.4
liger_kernel==0.5.5
livecc-utils==0.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmhash3==3.0.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.2.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
none==0.1.1
notebook==7.4.1
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnx==1.17.0
onnxruntime-gpu==1.21.0
openai==1.68.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
overrides==7.7.0
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pathlib2==2.3.7.post1
pathtools==0.1.2
peft==0.15.0
pexpect==4.9.0
pillow==10.2.0
pillow_heif==0.22.0
platformdirs==4.3.7
pluggy==1.5.0
ply==3.11
prettytable==3.15.1
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==11.0.0
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.0.9
PySocks==1.7.1
pysubs2==1.8.0
pytest==6.2.5
python-apt==2.6.0
python-consul==1.1.0
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-engineio==4.11.2
python-etcd==0.4.5
python-json-logger==3.3.0
python-magic==0.4.26
python-multipart==0.0.20
python-socketio==5.12.1
python-speech-features==0.6
pytz==2022.5
pyxdg==0.28
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.10
redis==4.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
schedule==1.2.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.24.0
setproctitle==1.3.5
shellingham==1.5.4
shortuuid==1.0.13
simple-websocket==1.1.0
simsimd==6.2.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==5.3.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-websupport==2.0.0
stack-data==0.6.3
starlette==0.46.1
stringzilla==3.12.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
thriftpy2==0.5.2
tifffile==2025.3.13
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tox==3.28.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.2.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
unattended-upgrades==0.1
unidiff==0.7.3
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
virtualenv==20.29.3
wadllib==1.3.6
watchdog==6.0.0
wavfile==4.7.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xdg==5
xxhash==3.5.0
yarl==1.18.3
yt-dlp==2025.2.19
zipp==3.21.0
zstandard==0.19.0
